shadow dom created on load and modals append shadow dom
